first steps towards supporting SSL over uTP. moved the TODO comment and made it describe the next step. Optimized session destruction in unit tests, hopefully shaving off a bit of test run-time (and now supports asio-debugging of unit tests)

+++ test/main.cpp
@@ -159,10 +159,14 @@ int main()
 	fflush(stdout);
 	fflush(stderr);
 
-	remove_all(test_dir, ec);
-	if (ec)
-		fprintf(stderr, "failed to remove test dir: %s\n", ec.message().c_str());
+	int ret = print_failures();
+	if (ret == 0)
+	{
+		remove_all(test_dir, ec);
+		if (ec)
+			fprintf(stderr, "failed to remove test dir: %s\n", ec.message().c_str());
+	}
 
-	return print_failures();
+	return ret;
 }
